+++ CHANGES.md
@@ -56,7 +56,7 @@
 
 ### _Blackd_
 
-<!-- Changes to blackd -->
+- Windows style (CRLF) newlines will be preserved (#3257).
 
 ### Integrations